r1228: use int64_t instead of long long
(jra: please use: void, char int, uint_t, [u]int<8|16|32|64>_t types in new code)

metze

+++ source/lib/time.c
@@ -410,9 +410,9 @@ NTTIME nttime_from_string(const char *s)
 	return strtoull(s, NULL, 0);
 }
 
-long long usec_time_diff(struct timeval *larget, struct timeval *smallt)
+int64_t usec_time_diff(struct timeval *larget, struct timeval *smallt)
 {
-	long long sec_diff = larget->tv_sec - smallt->tv_sec;
-	return (sec_diff * 1000000) + (long long)(larget->tv_usec - smallt->tv_usec);
+	int64_t sec_diff = larget->tv_sec - smallt->tv_sec;
+	return (sec_diff * 1000000) + (int64_t)(larget->tv_usec - smallt->tv_usec);
 }
